requirements.txt
quart
boto3
langchain
langchain-community
botocore
python-dotenv
